+++ src/VerifierAlgorithm.ts
@@ -167,3 +167,4 @@ function VerifierAlgorithm(alg: string): Verifier {
 VerifierAlgorithm.toSignatureObject = toSignatureObject
 
 export default VerifierAlgorithm
+
